requirements.txt
django-scheduler
appnope==0.1.0
decorator==4.0.10
numpy==1.11.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-speech-features==0.4
scikit-learn==0.18.1
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
traitlets==4.3.1
wcwidth==0.1.7
